[AIP-21] Move contrib.sensors.bash_sensor sensors.bash
