+++ scripts/ci/pylint_todo.txt
@@ -75,7 +75,7 @@
 ./airflow/contrib/operators/winrm_operator.py
 ./airflow/providers/amazon/aws/sensors/glue_catalog_partition.py
 ./airflow/providers/microsoft/azure/sensors/azure_cosmos.py
-./airflow/contrib/sensors/bash_sensor.py
+./airflow/sensors/bash.py
 ./airflow/contrib/sensors/celery_queue_sensor.py
 ./airflow/contrib/sensors/datadog_sensor.py
 ./airflow/providers/amazon/aws/sensors/emr_base.py
